JWT Authentication

+++ backend/controllers/goalController.js
@@ -1,4 +1,6 @@
 const asyncHandler = require('express-async-handler')
+
+const User = require('../models/userModel')
 const Goal = require('../models/goalModel')
 
 // @desc    Get goals
@@ -12,7 +14,8 @@ const Goal = require('../models/goalModel')
 //     res.status(200).json({ message: 'Get goals' })
 // })
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    // const goals = await Goal.find({})
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -39,6 +42,20 @@ const getGoals = asyncHandler(async (req, res) => {
 
 //     res.status(200).json({ message: 'Set goal' })
 // })
+
+// const setGoal = asyncHandler(async (req, res) => {
+//     if (!req.body.text) {
+//         res.status(400)
+//         throw new Error('Please add a text field')
+//     }
+
+//     const goal = await Goal.create({
+//         text: req.body.text,
+//     })
+
+//     res.status(200).json(goal)
+// })
+
 const setGoal = asyncHandler(async (req, res) => {
     if (!req.body.text) {
         res.status(400)
@@ -47,6 +64,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -62,11 +80,40 @@ const setGoal = asyncHandler(async (req, res) => {
 // const updateGoal = asyncHandler(async (req, res) => {
 //     res.status(200).json({ message: `Update goal ${req.params.id}` })
 // })
+
+// const updateGoal = asyncHandler(async (req, res) => {
+//     const goal = await Goal.findById(req.params.id)
+//     if (!goal) {
+//         res.status(400)
+//         throw new Error('Goal not found')
+//     }
+
+//     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+//         new: true,
+//     })
+
+//     res.status(200).json(updatedGoal)
+// })
+
 const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -87,12 +134,41 @@ const updateGoal = asyncHandler(async (req, res) => {
 //     res.status(200).json({ message: `Delete goal ${req.params.id}` })
 // })
 
+// const deleteGoal = asyncHandler(async (req, res) => {
+//     const goal = await Goal.findById(req.params.id)
+//     if (!goal) {
+//         res.status(400)
+//         throw new Error('Goal not found')
+//     }
+//     // remove is not a function in mongoose
+//     // await goal.remove()
+//     await goal.deleteOne()
+
+//     res.status(200).json({ id: req.params.id })
+// })
+
+
 const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+    
     // remove is not a function in mongoose
     // await goal.remove()
     await goal.deleteOne()
